refactor: reduce extra component file and extra assets folder
